02_LLM/gpt_api_tutorial/requirements.txt
python-dotenv==1.0.0
openai==1.3.0 
flask==3.1.0